handle ssh urls that use home relative paths (ie. paths starting with ~)
It's common for us to have scp style remote urls that use the '~' at the start of the path, to indicate the home directory of the user used at the start of the url. The conversion of these scp style urls to ssh:// urls breaks the path and adds an initial '/' to the `deploy_to` path, causing git-deploy not to work.

There is no way (that I know of) to preserve these paths when converting to an ssh:// style uri, so we need to detect this situation and 'correct' the url by removing the initial '/' when getting the path