Don't display an error when no db/migrate
before_restart.rb would make git print an error message when run in a
Rails project without db/migrate. The error would be printed to stderr
which would result in num_migrations being set to 0 (which is correct).
However, printing the error message when there was no error is
confusing.

This commit introduces a check for the existence of db/migrate. If it
exists then git is called. Otherwise num_migrations is set to 0.